added floating-point calculation

+++ main.py
@@ -8,12 +8,12 @@ def main():
         print("Error! Not enough arguments")
         print(USAGE)
     elif len(sys.argv) == 2:
-        print(9 / 5 * int(sys.argv[1]) + 32)
+        print(9 / 5 * float(sys.argv[1]) + 32)
     elif len(sys.argv) == 3:
         if sys.argv[1] == '-C':
-            print(9 / 5 * int(sys.argv[2]) + 32)
+            print(9 / 5 * float(sys.argv[2]) + 32)
         elif sys.argv[1] == '-F':
-            print(5 / 9 * (int(sys.argv[2]) - 32))
+            print(5 / 9 * (float(sys.argv[2]) - 32))
         else:
             print("Error! Unsupported scale")
             print(USAGE)
